+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Change Log
 
+## [0.2.0] - 2023-04-06
+### Added
+- <code>&dollar;{command:<em>name</em>}</code> and <code>&dollar;{command:cv#<em>name</em>}</code>
+- variable substitution on picked values
+
 ## [0.1.0] - 2023-03-26
 - First release